requirements.txt
asgiref==3.3.0
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.3
djangorestframework==3.12.1
Faker==4.18.0
gunicorn==20.0.4
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
model-bakery==1.2.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.7.2
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
wrapt==1.12.1
